# lab1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для расчета должной жизненной емкости легких (ДЖЕЛ)\nЛабораторная работа 4 - Добавление аутентификации и авторизации",
        "title": "Lung Capacity Calculation API",
        "contact": {
            "name": "API Support",
            "url": "http://localhost:8080",
            "email": "support@lungcalc.ru"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/pvlc-med-cards/{id}/async-result": {
            "put": {
                "description": "Обновляет результаты ДЖЕЛ после асинхронного расчета в Django сервисе\nТребует правильный ключ авторизации в поле async_key",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "medical-cards"
                ],
                "summary": "Обновление результатов асинхронного расчета",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Результаты расчета от Django сервиса",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.AsyncResultUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.AsyncResultUpdateRequest": {
            "type": "object",
            "required": [
                "async_key",
                "total_result"
            ],
            "properties": {
                "async_calculated": {
                    "description": "AsyncCalculated - флаг что расчет асинхронный",
                    "type": "boolean"
                },
                "async_key": {
                    "description": "AsyncKey - ключ для авторизации (ОБЯЗАТЕЛЬНОЕ ПОЛЕ!)\nDjango должен отправлять тот же ключ что и в константе AsyncServiceKey",
                    "type": "string"
                },
                "calculated_count": {
                    "description": "CalculatedCount - сколько формул рассчитано",
                    "type": "integer"
                },
                "individual_results": {
                    "description": "==================== ДОБАВЛЯЕМ ПОЛЕ ДЛЯ ИНДИВИДУАЛЬНЫХ РЕЗУЛЬТАТОВ ====================\nIndividualResults - результаты для каждого расчета (пациента)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.IndividualResult"
                    }
                },
                "total_result": {
                    "description": "TotalResult - результат расчета ДЖЕЛ от Django",
                    "type": "number"
                }
            }
        },
...
        "ds.IndividualResult": {
            "type": "object",
            "properties": {
                "formula_id": {
                    "type": "integer"
                },
                "individual_result": {
                    "type": "number"
                },
                "input_height": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT токен для аутентификации. Формат: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}